split security config to remove circular dependencies
