+++ backend/src/main/java/gov/cdc/usds/simplereport/config/DevCorsSecurityConfiguration.java
@@ -1,0 +1,22 @@
+package gov.cdc.usds.simplereport.config;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+/** Stub no-op configuration for development and test environments. */
+@Profile(BeanProfiles.NO_SECURITY)
+@Configuration
+@ConditionalOnWebApplication
+@Slf4j
+public class DevCorsSecurityConfiguration implements WebMvcConfigurer {
+
+  @Override
+  public void addCorsMappings(CorsRegistry registry) {
+    log.warn("CORS ENABLED BY {} PROFILE", BeanProfiles.NO_SECURITY);
+    registry.addMapping("/**").allowedMethods("*");
+  }
+}
