+++ backend/src/main/java/gov/cdc/usds/simplereport/config/OktaLocalSecurityConfiguration.java
@@ -3,9 +3,7 @@ package gov.cdc.usds.simplereport.config;
 import com.okta.spring.boot.oauth.Okta;
 import gov.cdc.usds.simplereport.service.model.IdentityAttributes;
 import gov.cdc.usds.simplereport.service.model.IdentitySupplier;
-import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.autoconfigure.security.servlet.EndpointRequest;
 import org.springframework.boot.actuate.health.HealthEndpoint;
 import org.springframework.boot.actuate.info.InfoEndpoint;
@@ -21,9 +19,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.web.servlet.config.annotation.CorsRegistration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 /**
  * Live (with Okta integration) request-level security configuration, but skips audit logging for
@@ -34,12 +29,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Profile("!" + BeanProfiles.NO_SECURITY + " & " + BeanProfiles.CREATE_SAMPLE_DEVICES)
 @ConditionalOnWebApplication
 @Slf4j
-public class OktaLocalSecurityConfiguration extends WebSecurityConfigurerAdapter
-    implements WebMvcConfigurer {
-
-  @Autowired CorsProperties _corsProperties;
-
-  public static final String SAVED_REQUEST_HEADER = "SPRING_SECURITY_SAVED_REQUEST";
+public class OktaLocalSecurityConfiguration extends WebSecurityConfigurerAdapter {
 
   public interface OktaAttributes {
     String EMAIL = "email";
@@ -146,20 +136,5 @@ public class OktaLocalSecurityConfiguration extends WebSecurityConfigurerAdapter
       throw new RuntimeException(
           "Unexpected authentication principal of type " + principal.getClass());
     };
-  }
-
-  @Override
-  public void addCorsMappings(CorsRegistry registry) {
-    CorsRegistration reg = registry.addMapping("/**");
-
-    List<String> methods = _corsProperties.getAllowedMethods();
-    if (methods != null && !methods.isEmpty()) {
-      reg.allowedMethods(methods.toArray(String[]::new));
-    }
-
-    List<String> origins = _corsProperties.getAllowedOrigins();
-    if (origins != null && !origins.isEmpty()) {
-      reg.allowedOrigins(origins.toArray(String[]::new));
-    }
   }
 }
